move gps files to gps folder
